Switched int32s to int64s

+++ tgtype/Contact.go
@@ -5,5 +5,5 @@ type Contact struct {
 	PhoneNumber string `json:"phone_number"`
 	FirstName   string `json:"first_name"`
 	LastName    string `json:"last_name,omitempty"`
-	UserID      int32  `json:"user_id,omitempty"`
+	UserID      int64  `json:"user_id,omitempty"`
 }
