+++ tgtype/GameHighScore.go
@@ -2,7 +2,7 @@ package tgtype
 
 // GameHighScore - This object represents one row of the high scores table for a game.
 type GameHighScore struct {
-	Position int32 `json:"position"`
+	Position int64 `json:"position"`
 	User     *User `json:"user"`
-	Score    int32 `json:"score"`
+	Score    int64 `json:"score"`
 }
