+++ tgtype/Message.go
@@ -2,16 +2,16 @@ package tgtype
 
 // Message - This object represents a message.
 type Message struct {
-	MessageID             int32           `json:"message_id"`
+	MessageID             int64           `json:"message_id"`
 	From                  *User           `json:"from,omitempty"`
-	Date                  int32           `json:"date"`
+	Date                  int64           `json:"date"`
 	Chat                  *Chat           `json:"chat"`
 	ForwardFrom           *User           `json:"forward_from,omitempty"`
 	ForwardFromChat       *Chat           `json:"forward_from_chat,omitempty"`
-	ForwardFromMessageID  int32           `json:"forward_from_message_id,omitempty"`
-	ForwardDate           int32           `json:"forward_date,omitempty"`
+	ForwardFromMessageID  int64           `json:"forward_from_message_id,omitempty"`
+	ForwardDate           int64           `json:"forward_date,omitempty"`
 	ReplyToMessage        *Message        `json:"reply_to_message,omitempty"`
-	EditDate              int32           `json:"edit_date,omitempty"`
+	EditDate              int64           `json:"edit_date,omitempty"`
 	Text                  string          `json:"text,omitempty"`
 	Entities              []MessageEntity `json:"entities,omitempty"`
 	Audio                 *Audio          `json:"audio,omitempty"`
@@ -33,7 +33,7 @@ type Message struct {
 	GroupChatCreated      bool            `json:"group_chat_created,omitempty"`
 	SupergroupChatCreated bool            `json:"supergroup_chat_created,omitempty"`
 	ChannelChatCreated    bool            `json:"channel_chat_created,omitempty"`
-	MigrateToChatID       int32           `json:"migrate_to_chat_id,omitempty"`
-	MigrateFromChatID     int32           `json:"migrate_from_chat_id,omitempty"`
+	MigrateToChatID       int64           `json:"migrate_to_chat_id,omitempty"`
+	MigrateFromChatID     int64           `json:"migrate_from_chat_id,omitempty"`
 	PinnedMessage         *Message        `json:"pinned_message,omitempty"`
 }
