+++ tgtype/User.go
@@ -2,7 +2,7 @@ package tgtype
 
 // User - This object represents a Telegram user or bot.
 type User struct {
-	ID        int32  `json:"id"`
+	ID        int64  `json:"id"`
 	FirstName string `json:"first_name"`
 	LastName  string `json:"last_name,omitempty"`
 	Username  string `json:"username,omitempty"`
